Time: 0 ms (100.00%), Space: 5.9 MB (26.56%) - LeetHub

+++ 1281-subtract-the-product-and-sum-of-digits-of-an-integer/1281-subtract-the-product-and-sum-of-digits-of-an-integer.cpp
@@ -1,7 +1,7 @@
 class Solution {
 public:
     int subtractProductAndSum(int n) {
-        long long int pro=1, sum=0;
+        int pro=1, sum=0;
         while(n>0)
         {
             int r=n%10;
